♻️ refactor(user): alter user model and split remote user types
- add comments in user class file to describe it as an Anti-Corruption Layer and domain model
- divide remote user types into individual files for each provider
- rename `remote-user.ts` to `remote-user-apple.ts`
